utils/requirements.txt
accelerate
av
numpy
Pillow
imageio
oci
git+https://github.com/huggingface/transformers
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.7.0
torchvision
torchaudio